Rename zabbix_host_get.py => zabbix_host_facts.py
